1.0.1 Update: Finished Standalone Mode for Wrapper and register process to master, added reconnect methods for Wrapper Netty Client, simplified ServerManager, fixed CloudPlugin for Bukkit, bugfixes in Server Classes, implemented Thread-Pool for all Wrapper Server instances, fixed ServerCountUpdatePacket (missing UUID) & WrapperRegisterFromStandalonePacket (missing server-name), added missing packets to Protocol, minor bugfixes & improvements
//ToDo: Fix BungeeCord ConsoleOutputStream, Implement Plugin command for reconnecting the plugin after MasterServer shutdown

+++ core/src/main/java/eu/mcone/cloud/core/network/packet/ServerInfoPacket.java
@@ -35,7 +35,7 @@ public class ServerInfoPacket extends Packet {
             out.writeUTF(serverInfo.getName());
             out.writeUTF(serverInfo.getTemplateName());
             out.writeInt(serverInfo.getTemplateID());
-            out.writeInt(serverInfo.getRam());
+            out.writeLong(serverInfo.getRam());
             out.writeInt(serverInfo.getPort());
             out.writeInt(serverInfo.getMaxPlayers());
             out.writeUTF(serverInfo.getVersion().toString());
@@ -59,7 +59,7 @@ public class ServerInfoPacket extends Packet {
             String name = input.readUTF();
             String templateName = input.readUTF();
             int templateId = input.readInt();
-            int ram = input.readInt();
+            long ram = input.readLong();
             int port = input.readInt();
             int maxplayers = input.readInt();
             ServerVersion version = ServerVersion.valueOf(input.readUTF());
